refactor metrics into subclasses
